refactor: fix db config
