+++ nest-rest-service/src/app.module.ts
@@ -5,9 +5,10 @@ import { DbModule } from './db/db.module';
 import { UsersModule } from './resources/users/users.module';
 import { BoardsModule } from './resources/boards/boards.module';
 import { TasksModule } from './resources/tasks/tasks.module';
+import { ConfigModule } from '@nestjs/config';
 
 @Module({
-  imports: [DbModule, UsersModule, BoardsModule, TasksModule],
+  imports: [DbModule, UsersModule, BoardsModule, TasksModule, ConfigModule.forRoot({ isGlobal: true })],
   controllers: [AppController],
   providers: [AppService],
 })
