+++ nest-rest-service/src/main.ts
@@ -2,8 +2,9 @@ import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 
 async function bootstrap() {
+  const { PORT } = process.env;
   const app = await NestFactory.create(AppModule);
-  await app.listen(4000);
+  await app.listen(PORT);
   console.log(`Application is running on: ${await app.getUrl()}`);
 }
 bootstrap();
